+++ src/data/historyitem.h
@@ -26,8 +26,8 @@ class HistoryItem : public AbstractDataItem
 {
 public:
     explicit HistoryItem();
-    int timestamp() const;
-    void setTimestamp(int timestamp);
+    long long timestamp() const;
+    void setTimestamp(long long timestamp);
     QByteArray favicon() const;
     void setFavicon(const QByteArray &favicon);
     int visitCount() const;
@@ -35,7 +35,7 @@ public:
     int loadingTime() const;
     void setLoadingTime(int loadingTime);
 private:
-    int m_timestamp;
+    long long m_timestamp;
     QByteArray m_favicon;
     int m_visitCount;
     int m_loadingTime;
